broke scenes file into multiple ts files
